requirements.txt
appdirs==1.4.3
Django==2.0.6
djangorestframework==3.8.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
uWSGI==2.0.15
psycopg2==2.7.3.2
model_mommy==1.6.0